+++ docs/CHANGELOG.md
@@ -5,6 +5,8 @@ CHANGELOG
 - New features:
   - Click anywhere on toolbar to open chapter selection view and move to current book (#206)
   - Add a seek bar to set font size scale (#206)
+- Changes:
+  - Make view pager for reading less sensitive to swipe
 - Update dependencies:
   - Gradle to 7.3
   - Kotlin to 1.6.0